lambda/pdf_generator/requirements.txt
# fpdf2 — pure-Python PDF generation (MIT licence, no system dependencies)
fpdf2>=2.7.9
